requirements.txt
asgiref==3.3.1
certifi==2020.11.8
chardet==3.0.4
click==7.1.2
Django==3.1.4
djangorestframework==3.12.2
easydict==1.9
Flask==1.1.2
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
PyMySQL==0.10.1
pytz==2020.4
qiniu==7.3.0
requests==2.25.0
sqlparse==0.4.1
urllib3==1.26.2
Werkzeug==1.0.1
wincertstore==0.2
